Test has --all option to test more cases.
