+++ tests/__init__.py
@@ -1,0 +1,1 @@
+#Now this is a module too
